creado un usercontrol para manejar los post que se van a crear
