+++ CapaVisual.csproj
@@ -99,11 +99,11 @@
     <Compile Include="LoginPage.Designer.cs">
       <DependentUpon>LoginPage.cs</DependentUpon>
     </Compile>
-    <Compile Include="MainPage.cs">
+    <Compile Include="AppWindow.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="MainPage.Designer.cs">
-      <DependentUpon>MainPage.cs</DependentUpon>
+    <Compile Include="AppWindow.Designer.cs">
+      <DependentUpon>AppWindow.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
@@ -113,17 +113,26 @@
     <Compile Include="RegisterPage.Designer.cs">
       <DependentUpon>RegisterPage.cs</DependentUpon>
     </Compile>
-    <Compile Include="MainPage2.cs">
+    <Compile Include="MainPage.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="MainPage2.Designer.cs">
-      <DependentUpon>MainPage2.cs</DependentUpon>
+    <Compile Include="MainPage.Designer.cs">
+      <DependentUpon>MainPage.cs</DependentUpon>
+    </Compile>
+    <Compile Include="PostCard.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="PostCard.Designer.cs">
+      <DependentUpon>PostCard.cs</DependentUpon>
     </Compile>
     <EmbeddedResource Include="LoginPage.resx">
       <DependentUpon>LoginPage.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="MainPage.resx">
-      <DependentUpon>MainPage.cs</DependentUpon>
+    <EmbeddedResource Include="AppWindow.resx">
+      <DependentUpon>AppWindow.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="PostCard.resx">
+      <DependentUpon>PostCard.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
@@ -138,8 +147,8 @@
     <EmbeddedResource Include="RegisterPage.resx">
       <DependentUpon>RegisterPage.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="MainPage2.resx">
-      <DependentUpon>MainPage2.cs</DependentUpon>
+    <EmbeddedResource Include="MainPage.resx">
+      <DependentUpon>MainPage.cs</DependentUpon>
     </EmbeddedResource>
     <None Include="packages.config" />
     <None Include="Properties\Settings.settings">
@@ -187,6 +196,30 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Uploads\" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Resources\favicon by iconSvg.co.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Resources\Be Like by iconSvg.co%281%29.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Resources\Be Like by iconSvg.co%282%29.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Resources\Dislike by iconSvg.co.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Resources\Messaging by iconSvg.co.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Resources\Share by iconSvg.co.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Resources\Share by iconSvg.co%282%29.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Resources\downytfytload.jpg" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
