+++ Program.cs
@@ -16,7 +16,7 @@ namespace CapaVisual
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new MainPage());
+            Application.Run(new AppWindow());
         }
     }
 }
